# parent/com.bike.server/target/generated-sources/openapi/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: jacquetq@gmail.com
  description: |
    Bike Rest Api with Spring Boot
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Bike Api
  version: 1.0.0
...
servers:
- description: SwaggerHub API Auto Mocking
  url: https://virtserver.swaggerhub.com/QuentinJacquet/com.bike.api/1.0.0
...
paths:
  /bike:
    get:
      operationId: getBikes
      parameters:
      - description: status of bikes
        explode: true
        in: query
        name: status
        required: false
        schema:
          enum:
          - available
          - used
          - under repair
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Bike'
                type: array
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/Bike'
                type: array
          description: successful operation
        "405":
          description: Invalid input
      summary: get bikes according to filter
      tags:
      - bike
      x-accepts: application/json
      x-tags:
      - tag: bike
    post:
      operationId: addBike
      requestBody:
        $ref: '#/components/requestBodies/BikeArray'
      responses:
        "200":
          description: successful operation
        "405":
          description: Invalid input
      summary: Add new bikes
      tags:
      - bike
      x-contentType: application/json
      x-accepts: application/json
      x-tags:
      - tag: bike
  /bike/{bikeId}:
    delete:
      operationId: deleteBike
      parameters:
      - description: Bike id to delete
        explode: false
        in: path
        name: bikeId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: Succesful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Bike not found
      summary: Deletes a bike
      tags:
      - bike
      x-accepts: application/json
      x-tags:
      - tag: bike
    get:
      description: Returns a single bike
      operationId: getBikeById
      parameters:
      - description: ID of bike to return
        explode: false
        in: path
        name: bikeId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bike'
            application/xml:
              schema:
                $ref: '#/components/schemas/Bike'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Bike not found
      summary: Find bike by ID
      tags:
      - bike
      x-accepts: application/json
      x-tags:
      - tag: bike
  /bike/{bikeId}/uploadImage:
    post:
      operationId: uploadFile
      parameters:
      - description: ID of bike to update
        explode: false
        in: path
        name: bikeId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/octet-stream:
            schema:
              format: binary
              type: string
      responses:
        "200":
          description: successful operation
      summary: uploads an image
      tags:
      - bike
      x-contentType: application/octet-stream
      x-accepts: application/json
      x-tags:
      - tag: bike
...
components:
  requestBodies:
...
    BikeArray:
      content:
        application/json:
          schema:
            items:
              $ref: '#/components/schemas/Bike'
            type: array
        application/xml:
          schema:
            items:
              $ref: '#/components/schemas/Bike'
            type: array
      description: List of bike object
      required: true
...
  schemas:
    Bike:
      example:
        photoUrls:
        - photoUrls
        - photoUrls
        id: 0
        status: available
      properties:
        id:
          format: int64
          readOnly: true
          type: integer
        status:
          default: available
          description: bike status
          enum:
          - available
          - used
          - under repair
          readOnly: true
          type: string
        photoUrls:
          items:
            type: string
          type: array
          xml:
            name: photoUrl
            wrapped: true
      type: object
      xml:
        name: Bike
...